+++ src/app/components/core/hoc/PageLoader.jsx
