--- src/app/constants/Platform.jsx
@@ -1,9 +1,0 @@
-const Platform = [
-  { id: 1, name: "GitHub", value: "github" },
-  { id: 2, name: "LinkedIn", value: "linkedin" },
-  { id: 3, name: "Behance", value: "behance" },
-  { id: 4, name: "Google", value: "google" },
-  { id: 5, name: "Facebook", value: "facebook" },
-]
-
-export default Platform;
