--- src/app/constants/TechStack.js
@@ -1,11 +1,0 @@
-const TechStack=[
-    {id:1,name:"Backend"},
-    {id:2,name:"Frontend"},
-    {id:3,name:"UXUI"},
-    {id:4,name:"QAQC"},
-    {id:5,name:"Full Stack"},
-    {id:6,name:"DevOps"}
-]
-
-
-export default TechStack;
