+++ src/app/constants/index.jsx
@@ -1,0 +1,22 @@
+const Platform = [
+  { id: 1, name: "GitHub", value: "github" },
+  { id: 2, name: "LinkedIn", value: "linkedin" },
+  { id: 3, name: "Behance", value: "behance" },
+  { id: 4, name: "Google", value: "google" },
+  { id: 5, name: "Facebook", value: "facebook" },
+]
+
+
+
+const TechStack=[
+    {id:1,name:"Backend"},
+    {id:2,name:"Frontend"},
+    {id:3,name:"UXUI"},
+    {id:4,name:"QAQC"},
+    {id:5,name:"Full Stack"},
+    {id:6,name:"DevOps"}
+]
+
+
+
+export default {Platform,TechStack};
