+++ src/app/features/auth/pages/OpomRegisterPage.jsx
@@ -5,8 +5,7 @@ import React, { useEffect, useState } from "react";
 import { Controller, useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
 import { authUtils } from "@/lib/utils";
-import TechStack from "../../../constants/TechStack";
-import Platform from "@/constants/Platform";
+import {TechStack,Platform} from "@/constants/index"
 import PhoneInput from "react-phone-input-2";
 import "react-phone-input-2/lib/style.css";
 import Button from "@/components/ui/Button";
